Expectation–Maximization (EM) Algorithm

The EM algorithm finds the maximum-likelihood estimates of a HMM.

Usage,

em(HM,X)

Arguments

HM

A HMM object

X

Data

tol

tolerance of percentage change between likelihoods from different iterations

maxiter

the maximum number of iteration the algorithm will run

Value

HMM Returns an updated HMM object

Details

Is an iterative method to find a local maximum likelihood or a maximum poosteriori estimates of parameters in a statistical model, where the model depends on hidden variables. The EM alternates between performing an expectation step, which creates a function for the expectation of the log-likelihood evaluated using the curretn estimate for the parameters, and a maximation step, which computes the parameters, which maximize the expected log-likelihood found in the E step. This is then iterated.

Examples

 X <- earthquakes$n
delta = c(0.5,0.5)
trans=matrix(c(0.9,0.1,0.1,0.9),2,2)
HM = HMM(stationary_dist = delta,transmission = trans,  
        emission_function_names = c("dpois","dpois"),parameters = list(list(lambda=10),list(lambda=30)) )
em(HM=HM,X=X)
#> [1] "Number of iterations: 21"
#> [1] "This is a Hidden Markov model with 2 hidden states."
#> [1] "It has the Statinonary distribution of:"
#> [1] 1.000000e+00 1.104646e-56
#> [1] "It has the transmision matrix:"
#>            [,1]     [,2]
#> [1,] 0.92837683 0.119045
#> [2,] 0.07162317 0.880955
#> [1] "State 1  has the emission function dpois with parameters lambda = 15.4205545371766"
#> [1] "State 2  has the emission function dpois with parameters lambda = 26.0175459459797"